marketing and features

+++ Assets/FeatureView2.cs
@@ -18,7 +18,7 @@ public class FeatureView2 : View
         Feature = newProductFeature;
 
         var competitors = Companies.GetDirectCompetitors(Flagship, Q, true);
-        var upgrades = competitors.Select(c => Visuals.Colorize(Products.GetFeatureRating(c, Feature.Name) + "", c.isFlagship ? Colors.COLOR_BEST : Colors.COLOR_NEUTRAL));
+        var upgrades = competitors.Select(c => Visuals.Colorize((int)Products.GetFeatureRating(c, Feature.Name) + "", c.isFlagship ? Colors.COLOR_BEST : Colors.COLOR_NEUTRAL));
 
         var space = "      ";
